Move out the .inf parsing into a new module
Also add a proper CLI and some unit tests.
