+++ .travis.yml
@@ -11,5 +11,6 @@ before_script:
 script:
   - docker run -t -v `pwd`:/build centos7 ./uploadedfile_test.py
   - docker run -t -v `pwd`:/build centos7 ./cabarchive/tests/archive_test.py
+  - docker run -t -v `pwd`:/build centos7 ./infparser/tests/parser_test.py
   - docker run -t -v `pwd`:/build centos7 ./pylint_test.py
   - docker run --privileged -t -v `pwd`:/build centos7 ./lvfs_test.py
