+++ infparser/__init__.py
@@ -1,0 +1,7 @@
+#!/usr/bin/python3
+# -*- coding: utf-8 -*-
+#
+# Copyright (C) 2018-2019 Richard Hughes <richard@hughsie.com>
+# Licensed under the GNU General Public License Version 2
+
+from . parser import InfParser
