+++ infparser/tests/firmware.inf
@@ -1,0 +1,20 @@
+; Copyright (C) 2015 Hughski Limited
+
+[Version]
+Class=Firmware
+ClassGuid={f2e7dd72-6468-4e36-b6f1-6488f42c1b52}
+DriverVer=04/18/2015,2.0.3
+
+[Firmware_CopyFiles]
+firmware.bin
+
+[Firmware_AddReg]
+HKR,,FirmwareId,,{2082b5e0-7a64-478a-b1b2-e3404fab6dad}
+HKR,,FirmwareVersion,%REG_DWORD%,0x0000000
+HKR,,FirmwareFilename,,firmware.bin
+
+[Strings]
+Provider     = "Hughski"
+MfgName      = "Hughski Limited"
+FirmwareDesc = "ColorHug2 Firmware"
+DiskName     = "Firmware for the ColorHug2 Colorimeter"
